+++ src/navigation/index.ts
@@ -1,2 +1,0 @@
-export { default as BottomTabNavigator } from './bottomTabNavigator'
-export { default as StackNavigator } from './stackNavigator'