+++ src/screens/LoginScreen.tsx
@@ -2,13 +2,15 @@ import i18n from '../i18n'
 import React, { useState } from 'react'
 import { StyleSheet, Text, Image } from 'react-native'
 import { Content, View } from 'native-base'
+import { AntDesign } from '@expo/vector-icons'
 import { TouchableOpacity } from 'react-native-gesture-handler'
 import { SharedElement } from 'react-native-shared-element'
 import { useNavigation } from '@react-navigation/native'
 // import Animated, { useCode } from 'react-native-reanimated'
 
+import { WelcomeScreenNavigationProps } from '../@types/navigation'
 import LoginForm from '../components/forms/LoginForm'
-import { GradientContainer } from '../components/atomic/'
+import { GradientContainer, StyledText } from '../components/atomic/'
 import { Colours, Fonts } from '../styles'
 import { SignInNavContext } from '../contexts'
 import { normalise, scaleImageByScreenDimensions } from '../../src/helpers/'
@@ -21,14 +23,16 @@ const LOGO_IMAGE_PATH = '../../assets/images/logo.png'
 export default function LoginScreen() {
   // const logoFontSize = useState(new Value(0))[0]
 
-  const nav = useNavigation()
-  const _transit = () => {
-    nav.goBack()
-  }
+  const nav = useNavigation<WelcomeScreenNavigationProps>()
 
   /* events triggered when `Sign Up` text link is clicked are defined here */
-  const _signUpTxtOnPressed = () => {
+  const _signInTxtOnPressed = () => {
     nav.navigate('SignUp')
+  }
+
+  const _goBack = () => {
+    /* see CHANGELOG of 01/09/2020 */
+    nav.navigate('Welcome')
   }
 
   /**@TODO on the first run, shrink the logo font size
@@ -42,6 +46,7 @@ export default function LoginScreen() {
   return (
     <GradientContainer flexDirection={'column'}>
       <Content scrollEnabled={false} contentContainerStyle={styles.contentContainer}>
+        {/* Logo */}
         <View style={styles.logoContainer}>
           <SharedElement id="logo">
             <Image
@@ -52,22 +57,45 @@ export default function LoginScreen() {
             />
           </SharedElement>
         </View>
+        {/* END Logo */}
+        {/* Account Identifier & Password input fields + Link to `ForgotPassword` screen */}
         <SignInNavContext.Provider value={nav}>
+          {/* Touchable 'Back' link */}
+          <View style={styles.backContainer}>
+            <TouchableOpacity
+              style={{ flexDirection: 'row' }}
+              onPress={_goBack}
+            >
+              <AntDesign
+                name="left"
+                size={normalise(16)}
+                color={Colours.White}
+                style={{ alignSelf: 'flex-start' }}
+              />
+              <StyledText fontWeight="bold">{i18n.t('signUp.backTxt')}</StyledText>
+            </TouchableOpacity>
+          </View>
+          {/* END Touchable 'Back' link */}
           <LoginForm />
         </SignInNavContext.Provider>
+        {/* END Account Identifier & Password input fields + Link to `ForgotPassword` screen */}
         <View style={styles.signUpTxtContainer}>
+          {/* `Have no account yet?` text */}
           <View>
             <Text style={[styles.forgetPasswordTxt, { textAlign: 'right' }]}>
               {`${i18n.t('signIn.askSignUpTxt')}`}
             </Text>
           </View>
+          {/* END `Have no account yet?` text */}
+          {/* Link to `SignUp` screen */}
           <TouchableOpacity
-            style={styles.signUpTouchableTxt}
+            style={styles.signInTouchableTxt}
             activeOpacity={0.6}
-            onPress={_signUpTxtOnPressed}
+            onPress={_signInTxtOnPressed}
           >
             <Text style={styles.signUpTxt}>{i18n.t('signIn.signUpTxt')}</Text>
           </TouchableOpacity>
+          {/* END Link to `SignUp` screen */}
         </View>
       </Content>
     </GradientContainer>
@@ -93,10 +121,19 @@ const styles = StyleSheet.create({
     ...scaleImageByScreenDimensions(require(LOGO_IMAGE_PATH), 0.8) 
   },
   logoContainer: {
-    flex: 1,
+    // PLEASE DO NOT EDIT THIS!
+    flex: 3,
     alignItems: 'center',
     justifyContent: 'center',
     paddingVertical: 15 + '%',
+  },
+  backContainer: {
+    flex: 0.5, // !! DANGEROUR ZONE FOR EDITTING !!
+    flexDirection: 'row',
+    width: 80 + '%',
+    paddingVertical: 16,
+    justifyContent: 'flex-start',
+    alignContent: 'flex-start',
   },
   signUpTxt: {
     flex: 2,
@@ -108,7 +145,7 @@ const styles = StyleSheet.create({
     textAlign: 'center',
     textAlignVertical: 'bottom',
   },
-  signUpTouchableTxt: {
+  signInTouchableTxt: {
     shadowOffset: {
       width: 1.25,
       height: 1.25,