+++ src/screens/SearchScreen.tsx
@@ -1,7 +1,9 @@
+import i18n from '../i18n'
 import React from 'react'
 import { StyleSheet, Text, View, Dimensions, TextInput } from 'react-native'
 import { TouchableOpacity } from 'react-native-gesture-handler'
 import { useNavigation } from '@react-navigation/native'
+import { SearchScreenNavigationProps } from '../@types/navigation'
 
 import GradientContainer from '../components/atomic/GradientContainer'
 import { HomeScreenNavigationProps } from '../@types/navigation'
@@ -14,19 +16,21 @@ const SCREEN_HEIGHT = Dimensions.get('window').height
 const SCREEN_WIDTH = Dimensions.get('window').width
 
 export default function SeacrhScreen() {
+  const navigation = useNavigation<SearchScreenNavigationProps>()
+
   return (
     <GradientContainer flexDirection={'column'}>
       <View style={styles.content}>
         <View style={styles.header}>
-          <Text style={styles.headerText}>Tra cuu</Text>
+          <Text style={styles.headerText}>{i18n.t('search._nav')}</Text>
         </View>
         <View style={styles.searchField}>
-          <TextInput style={styles.inputField} placeholder="So dien thoai" />
+          <TextInput style={styles.inputField} placeholder={i18n.t('search.phoneNumInput')} />
 
-          <TextInput style={styles.inputField} placeholder="Ma OTP" />
+          <TextInput style={styles.inputField} placeholder={i18n.t('search.otpCodeInput')} />
 
-          <TouchableOpacity style={styles.searchButton}>
-            <Text>Tra cuu</Text>
+          <TouchableOpacity style={styles.searchButton} onPress={() => navigation.navigate("SearchResult", {phone: undefined})}>
+            <Text>{i18n.t('search.submitBtn')}</Text>
           </TouchableOpacity>
         </View>
       </View>