+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # CHANGELOG 
 
+## Log: 01/09/2020
+
+Fixes based on PM's feedback:
+
+- [x] `LoginScreen`: Form quên mật khẩu, và nút quay lại màn hình giới thiệu
+- [x] `LoginScreen`: Nút quay lại dẫn đến màn hình giới thiệu
+- [ ] `SignUpScreen`: Đăng ký xong sẽ process đến màn hình đăng nhập chứ không process đến màn hình chính luôn
+
+Still in negotiation: 
+
+- [ ] `SearchScreen`: Sau khi tra cứu thì sẽ quay lại màn hình lịch sử tra cứu và tại cái lịch sử tra cứu đấy sẽ là hiển thị thông tin mới của số điện thoại vừa tra cứu (Phần này liên quan đến backend nên có thể tính sau nhưng process cứ phải chuẩn đã)
+
 ## Log: End of Week 8
 
 - [x] Redesign layout for `Welcome` screen
-- [ ] Fix bug: changing font family of 'user agreement' text components would interpolate newline character to them
-- [ ] Comeback touchable text for `SignUp` screen
+- [x] Fix bug: changing font family of 'user agreement' text components would interpolate newline character to them
+- [x] Comeback touchable text for `SignUp` screen
 - [ ] Form validation